Fix ambiguous String(float, uint8_t) overload
ESP32 core's WString.h has several explicit String(..., unsigned
char) constructors alongside String(float, unsigned int) -- passing
a uint8_t decimalPlaces directly is ambiguous (neither overload is a
strictly better match across both arguments). Caught at compile time
while flashing the batcave node; log()'s call site now casts to
unsigned int, and BME280Node's shared precision constants are
declared unsigned int to begin with.

+++ examples/BME280Node/BME280Node.ino
@@ -54,10 +54,15 @@ const std::vector<SensorNodeChannel> kChannels = {
 // up: BME280 is +-0.5C / +-3%RH / +-1hPa (temp/dew point/humidity/pressure), MAX17048 is
 // roughly +-1% SOC. Defined once and shared by both the log() call and the OLED text below, so
 // the two can't drift apart the way a second hardcoded printf precision could.
-const uint8_t kTempPrecision = 1;
-const uint8_t kHumidityPrecision = 0;
-const uint8_t kPressurePrecision = 1;
-const uint8_t kSocPrecision = 0;
+// unsigned int, not uint8_t: matches ESP32 core's String(float, unsigned int) constructor
+// exactly for the OLED lines below -- a uint8_t here is ambiguous against WString.h's other
+// explicit String(..., unsigned char) overloads (ties on "one exact arg, one converted arg" in
+// opposite positions). SensorNodeReading::decimalPlaces still narrows this to uint8_t for the
+// log() calls, which is a plain (non-overloaded) implicit conversion, so that's fine as-is.
+const unsigned int kTempPrecision = 1;
+const unsigned int kHumidityPrecision = 0;
+const unsigned int kPressurePrecision = 1;
+const unsigned int kSocPrecision = 0;
 
 SensorNode node;
 BME280 bme;
